broke MainController by trying to narrow down query results -- fix that

+++ src/main/java/hello/UserRepository.java
@@ -5,5 +5,6 @@ import hello.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
+
 }
